requirements.txt
numpy
pickle
pillow
sklearn
tensorflow
